Decode `&lt;` HTML entities
Due to parsing restrictions in Scaladoc(/Javadoc?) when documenting code
that uses HTML or XML you often need to encode tag opening angle
brackets (`<`) as `&lt;` so that it is not parsed as HTML elements in
the browser and not displayed.

However, sbt-doctest parses the text differently causing test failures
unless you rewrite `&lt;` as `<`, which in-turn causes browser display
problems.

This commit gives the best of both worlds by turning all occurrences of
`&lt;` back into `<` for the purposes of the doctests only.