Optionally decode HTML entities
Due to parsing restrictions in Scaladoc(/Javadoc?) when documenting code
that uses HTML or XML you often need to encode tag angle brackets (e.g.
`<` as `&lt;)` so that it is not parsed as HTML elements by
the browser and hidden.

However, sbt-doctest parses the text as-is causing test failures
unless you rewrite `&lt;` as `<`, which in-turn causes browser display
problems (back to problem one).

This commit gives the best of both worlds by creating an optional SBT
setting that decodes HTML entities in the doctests only.