Use fully qualified names in generated code
This prevents clashes with names that are defined in projects
that use sbt-doctest. A recent example of a name clash has been
observed in the libra project (https://github.com/to-ithaca/libra/issues/43)
where our use of `Unit` (which we assumed to be `_root_.scala.Unit`)
clashed with the library defined `libra.Unit` type.