Fix markdown rendering in readme
Currently there's a part of the readme that looks like this:

Any code blocks that start with the ````scala` markdown directive will be parsed.
It searches `*.md` under `baseDirectory` by default. It can be configured by
`doctestMarkdownPathFinder`.


The syntax is a little funky, but this PR changes it to look like this:

Any code blocks that start with the ```` ```scala```` markdown directive will be parsed.
It searches `*.md` under `baseDirectory` by default. It can be configured by
`doctestMarkdownPathFinder`.